// src/app/models/user.ts
export class User {
  _id?: string;
  firstName?: string;
  lastName?: string;
  address?: any;
  city?: string;
  state?: string;
  zip?: number;
  depaulID?: number;
  degree?: string;
  email?: string;
  userID?: string;
  password?: string;
  userType?: string;
  coursesTaken?: Array<any>;
  constructor() { }
}